Added Link for ClojureCLR repo
According to their GitHub Wiki - "ClojureCLR development closely tracks progress in the ClojureJVM project. We index many of our commits directly to commits in the Clojure repo, so it should be easy to track our progress. We try to be within a week or so of development milestones on the main Clojure project."

So I think we should not use alpha for it now and encourage everybody to give ClojureCLR a spin.